+++ src/components/RegistId/const.ts
