+++ src/components/RegistId/type.ts
