+++ src/components/RegisterInfo/RegisterInfo.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function RegistInfo() {
+  return <div>RegistInfo</div>;
+}
+
+export default RegistInfo;
