+++ src/pages/JoinPage/JoinPage.tsx
@@ -1,25 +1,39 @@
-import { useState } from 'react';
+import { useMemo, useState } from 'react';
 
-import LoginForm from '@src/components/LoginForm';
-import { LoginType } from '@src/components/LoginForm/type';
 import PageLayout from '@src/components/shared/PageLayout';
 import { HeaderPage } from '@src/components/Header/const';
+import RegistId from '@src/components/RegistId';
+import { LoginType } from '@src/components/RegistId/type';
+import RegistInfo from '@src/components/RegisterInfo';
 
 function JoinPage() {
-  const [profileId, setProfileId] = useState('');
+  const [step, setStep] = useState<number>(0);
+  const [profileId, setProfileId] = useState<string>('');
 
-  const handleClickLogin = async () => {
-    console.log(profileId);
+  const validateId = useMemo(() => {
+    if (profileId === '') return false;
+    return true;
+  }, [profileId]);
+
+  const toNextStep = () => {
+    if (step === 0) {
+      if (!validateId) return; // to do : Input invalid 표시
+    }
+    setStep(step + 1);
   };
 
   return (
     <PageLayout padding="20px" pageType={HeaderPage.JOIN} onlyContents>
-      <LoginForm
-        type={LoginType.KLIP}
-        value={profileId}
-        onClickLoginButton={handleClickLogin}
-        onChangeInput={setProfileId}
-      />
+      {step === 0 && (
+        <RegistId
+          type={LoginType.KLIP}
+          value={profileId}
+          handleClickButton={toNextStep}
+          onChangeInput={setProfileId}
+        />
+      )}
+
+      {step === 1 && <RegistInfo />}
     </PageLayout>
   );
 }
